separate data load from data prep
