+++ predict_prices/main.py
@@ -6,16 +6,16 @@ import pandas as pd
 from sklearn import ensemble, linear_model  # type: ignore
 import xgboost
 
+import data_load
 import data_prep
 from hyperparameter_tuning import ModelHyperparameterMap
-from data_prep import RawTestingData
 
 
 @dataclass
 class PurePipeline:
     model_hyperparameter_map: ModelHyperparameterMap
     train_data_prep_inputs: data_prep.TrainDataPrepInputs
-    testing_data: RawTestingData
+    testing_data: data_load.RawTestingData
     n_iter: int
     folds: int
 
@@ -89,9 +89,9 @@ def main():
     hyper_param_map = ModelHyperparameterMap(model_dict)
 
     # 1. read in train and test data
-    train_df = data_prep.load_training_data(train_filepath)
+    train_df = data_load.load_training_data(train_filepath)
 
-    test_df = data_prep.load_testing_data(test_filepath)
+    test_df = data_load.load_testing_data(test_filepath)
     train_data_prep_inputs = data_prep.TrainDataPrepInputs(train_df, ohe_bool, target_variable)
     pure_pipeline_inputs = PurePipeline(hyper_param_map,train_data_prep_inputs, test_df, n_iter, folds)
